pyproject.toml
[project]
name = "abschlussprojekt"
version = "0.1.0"
description = "DiscordBot"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "discord-py (>=2.5.2,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "openai (>=1.70.0,<2.0.0)",
    "scipy (>=1.15.2,<2.0.0)",
    "soundfile (>=0.13.1,<0.14.0)",
    "pydub (>=0.25.1,<0.26.0)",
    "psola (>=0.0.1,<0.0.2)",
    "llvmlite (==0.44.0)",
    "librosa (>=0.11.0,<0.12.0)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "yt-dlp (>=2025.3.31,<2026.0.0)",
    "ffmpeg (>=1.4,<2.0)",
    "pynacl (>=1.5.0,<2.0.0)",
    "ffmpeg-python (>=0.2.0,<0.3.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "spotipy (>=2.25.1,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"